requirements.txt
uwsgi
psycopg2-binary
Django==1.11.17
djangorestframework==3.8.2
ipython
django-import-export==1.1.0
ipdb
